backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
python-jose==3.3.0
pydantic==1.10.19
